+++ src/widgets/TbButton.php
@@ -138,6 +138,20 @@ class TbButton extends CWidget
 	 */
 	public $visible = true;
 
+    /**
+     * Tooltip for button
+     * @var bool
+     * @since 2.0.0
+     */
+    public $tooltip = false;
+
+    /**
+     * Tooltip options
+     * @var array
+     * @since 2.0.0
+     */
+    public $tooltipOptions = array();
+
 	/**
 	 *### .init()
 	 *
@@ -248,6 +262,25 @@ class TbButton extends CWidget
 		if (isset($this->completeText)) {
 			$this->htmlOptions['data-complete-text'] = $this->completeText;
 		}
+
+        if (isset($this->tooltip) && !$this->toggle) {
+            if (!is_array($this->tooltipOptions)) {
+                $this->tooltipOptions = array();
+            }
+
+            $this->htmlOptions['data-toggle'] = 'tooltip';
+            foreach ($this->tooltipOptions as $key => $value) {
+                $this->htmlOptions['data-' . $key] = $value;
+            }
+
+            /**
+             * Encode delay option
+             * @link http://getbootstrap.com/2.3.2/javascript.html#tooltips
+             */
+            if (isset($this->htmlOptions['data-delay']) && is_array($this->htmlOptions['data-delay'])) {
+                $this->htmlOptions['data-delay'] = CJSON::encode($this->htmlOptions['data-delay']);
+            }
+        }
 	}
 
 	/**